original flow maintained

+++ app/controllers/user_tags_controller.rb
@@ -61,8 +61,8 @@ class UserTagsController < ApplicationController
       if(!@user_tag.nil?)
           @user_tag = @user_tag.first 
       end 
-      if current_user.role == 'admin' || (!@user_tag.nil? && @user_tag.user == current_user)
-        if @user_tag
+      if current_user.role == 'admin' || params[:id] == current_user.id
+        if (!@user_tag.nil? && @user_tag.user == current_user)
           UserTag.where(uid: params[:id] , value: params[:name]).destroy_all    
           message = I18n.t('user_tags_controller.tag_deleted')
           output[:status] = true
